Did enough to stop the exception horror. Solution needs rethinking however.
The exception thrown when we couldn't find the server was dropping right
through. Not only this but we were returning a null string pointer which we
then tried to display. Ideally we should not be returning strings from
functions other than accessor functions. We should only return error codes.

I resolved the situation by catching the exception higher up (right in the
activity). However, I think I would prefer to catch it alot closer to where it
originates, and then return an error code.

e.g. Catch exception in the sendMessage() function itself, or in the send()
function. The send() function can then return an error enum (e.g.
DSGServerNotFound). The function that called msg.send() can then respond
appropriately.

Either way, we need to give this matter serious thought.