We now pass a username string to the GamesList Activity when we launch it.
After thinking about how we should handle passing User details around, I
initially thought that we should create an instance of the User class once we
have a successful login. We would then pass this to the GamesList activity.
However this is not only more complicated in terms of code (the User class
would have to implement the Parcelable interface), but it would be more
expensive in terms of resources. It is also unnecessary since I plan to store
all user data in a database that can be accessed from the GamesList Activity.

Thus I elected to simply pass the username string which we successfully logged
in with. This can be then used to access the database, retrieve game data and
session cookies, etc. from within the GamesList Activity. If we truly have need
for a User class at all, we can create the instance there. Note that this
strategy will require us to add the username, session cookie and (potentially)
password to the database once we login successfully in the Login Activity, so
that this information can be accessed from the GamesList Activity later on.