Fixed database and started developing Game class further.
Several fixes to resolve runtime issues with the database. We can now log in and check status. Started Developing Game class further to parse the status and get useful data to show to the user. Much more to do in the game class, but we are not far away from having a list of current games up and running.